# openapi.yaml
openapi: 3.0.3
info:
  title: Group Sessions Platform API
  version: 1.1.0
  description: |
    Initial OpenAPI specification for the Group Sessions Platform MVP.
    This specification is based on the ER diagram and covers the main
    REST operations for authentication, projects, rooms, participants,
    chat, board, comments and pomodoro management.
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                required: [ success ]
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailExists:
                  value:
                    error: EMAIL_ALREADY_EXISTS
                    message: User with this email already exists
                    details: null
                usernameExists:
                  value:
                    error: USERNAME_ALREADY_EXISTS
                    message: User with this username already exists
                    details: null
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                error: VALIDATION_ERROR
                message: Request validation failed
                details:
                  - field: email
                    message: field required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INTERNAL_SERVER_ERROR
                message: An unexpected error occurred on the server
                details: null
  /api/v1/auth/login:
    post:
      tags: [ auth ]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [ username, password ]
              properties:
                username:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: secret123
      responses:
        '200':
          description: Successful authentication
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly cookie with refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INVALID_CREDENTIALS
                message: Invalid email or password
                details: null
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                error: VALIDATION_ERROR
                message: Request validation failed
                details:
                  - field: password
                    message: field required
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INTERNAL_SERVER_ERROR
                message: An unexpected error occurred on the server
                details: null
  /api/v1/auth/refresh:
    post:
      tags: [ auth ]
      summary: Refresh JWT tokens
      operationId: refreshTokens
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Updated HttpOnly cookie with refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [ auth ]
      summary: Logout user
      operationId: logoutUser
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful logout
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears refresh token cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
...
components:
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v1/auth/login
          refreshUrl: /api/v1/auth/refresh
          scopes:
            users:read: Read user data
            projects:read: Read projects
            projects:write: Create and update projects
            tags:read: Read tags
            tags:write: Manage tags
            rooms:read: Read rooms
            rooms:write: Create, update and delete rooms
            participants:read: Read room participants
            participants:write: Manage room participants
            chat:read: Read room chat messages
            chat:write: Send, edit and delete messages
            board:read: Read board elements and comments
            board:write: Manage board elements and comments
            pomodoro:read: Read pomodoro state
            pomodoro:write: Control pomodoro timer
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: PROJECT_NOT_FOUND
        message:
          type: string
          example: Project with this id does not exist
        details:
          nullable: true
          oneOf:
            - type: object
            - type: string
      required: [error, message]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: field required
      required: [error, message, details]
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: alex123
        password:
          type: string
          format: password
          minLength: 6
          example: secret123
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJ...
        refresh_token:
          type: string
          example: eyJ...
        token_type:
          type: string
          example: bearer
      required: [ access_token, refresh_token, token_type ]
...